requirements.txt
jumbodb
pandas
numpy
spacy==3.0.6
torch~=1.9.0
transformers==4.9.2
spacy-transformers==1.0.4
spacy-canonicalizer
datasets==1.6.2
sentencepiece
protobuf
ujson
nltk
pexpect
anytree
sentence-transformers
plac
gdown
rank-bm25
pint==0.19.2
c3_NumParse
dateparser
tabulate
Levenshtein
elasticsearch==7.15.2